pyproject.toml
[project]
name = "sinew-and-steel"
version = "0.3.1"
description = "Agent-friendly tabletop RPG engine and tooling"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
]

[tool.uv]
# No dev dependencies yet.
